# openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Minimarket Plus API",
    "description": "API REST del sistema backend Minimarket Plus. Gestiona productos, categorías, carritos, inventario, ventas y usuarios con roles (cliente, cajero, administrador). Todos los endpoints de negocio requieren autenticación HTTP Basic; use el botón Authorize (credenciales de demostración en el README).",
    "contact": {
      "name": "Equipo Minimarket Plus - PBY2202",
      "email": "karla.santigut@gmail.com"
    },
    "license": {
      "name": "Uso académico - Duoc UC"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Entorno local de desarrollo"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/api/categorias/{id}": {
      "get": {
        "tags": [
          "Categorías"
        ],
        "summary": "Obtener una categoría por ID",
        "description": "Busca una categoría específica por su identificador único.",
        "operationId": "obtenerCategoriaPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la categoría",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "200": {
            "description": "Categoría encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          },
          "404": {
            "description": "No existe una categoría con el ID indicado"
          }
        }
      },
      "put": {
        "tags": [
          "Categorías"
        ],
        "summary": "Actualizar una categoría existente",
        "description": "Reemplaza los datos de la categoría identificada por el ID.",
        "operationId": "actualizarCategoria",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la categoría a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Nuevos datos de la categoría",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              },
              "examples": {
                "categoriaActualizada": {
                  "description": "categoriaActualizada",
                  "value": {
                    "nombre": "Bebidas"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "200": {
            "description": "Categoría actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          },
          "404": {
            "description": "No existe una categoría con el ID indicado"
          }
        }
      },
      "delete": {
        "tags": [
          "Categorías"
        ],
        "summary": "Eliminar una categoría",
        "description": "Elimina la categoría identificada por el ID. Sus productos asociados también se eliminan (cascade).",
        "operationId": "eliminarCategoria",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la categoría a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe una categoría con el ID indicado"
          },
          "204": {
            "description": "Categoría eliminada correctamente"
          }
        }
      }
    },
...
    "/api/categorias": {
      "get": {
        "tags": [
          "Categorías"
        ],
        "summary": "Listar todas las categorías",
        "description": "Retorna todas las categorías registradas para clasificar productos.",
        "operationId": "listarCategorias",
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "200": {
            "description": "Listado obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categoria"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorías"
        ],
        "summary": "Crear una nueva categoría",
        "description": "Registra una categoría con nombre único.",
        "operationId": "guardarCategoria",
        "requestBody": {
          "description": "Datos de la categoría a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              },
              "examples": {
                "nuevaCategoria": {
                  "description": "nuevaCategoria",
                  "value": {
                    "nombre": "Bebidas"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "201": {
            "description": "Categoría creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único de la categoría",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre único de la categoría",
            "example": "Bebidas"
          }
        },
        "description": "Categoría para clasificar los productos del minimarket"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "Autenticación HTTP Basic contra los usuarios registrados en H2",
        "name": "basicAuth",
        "scheme": "basic"
      }
    }
  }
}